# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "agentsfleet Control Plane API",
    "description": "API for managing workspaces, fleets, triggers, and runs.",
    "contact": {
      "name": "agentsfleet",
      "url": "https://agentsfleet.net"
    },
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://api.agentsfleet.net",
      "description": "Production"
    }
  ],
  "paths": {
...
    "/v1/tenants/me/workspaces": {
      "get": {
        "tags": [
          "Workspaces"
        ],
        "summary": "List the tenant's workspaces",
        "description": "Returns a stable oldest-first cursor page of workspaces owned by the caller's authoritative tenant. Pass `starting_after` from `next_cursor` to continue. The optional `name` filter uses exact equality and supports reconciliation after an uncertain workspace- create response. ",
        "operationId": "list_tenant_workspaces",
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "description": "Exact tenant-scoped workspace name.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "starting_after",
            "in": "query",
            "description": "Cursor returned as `next_cursor` by the previous page.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "How many rows to return.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspacesResponse"
                }
              }
            }
          },
          "401": {
            "description": "No credential, or one this route does not accept"
          },
          "403": {
            "description": "The credential is good and lacks the capability this route requires"
          },
          "500": {
            "description": "The daemon failed to answer"
          }
        }
      }
    },
...
    "/v1/workspaces": {
      "post": {
        "tags": [
          "Workspaces"
        ],
        "summary": "Create a workspace",
        "description": "Creates a named workspace in the caller's tenant. The server assigns the workspace identifier. This operation does not accept a replay key or retry automatically. After an uncertain response, query the tenant's workspaces with the exact name. Retrying the same tenant- unique name cannot create a second row and returns 409 when the first request committed. ",
        "operationId": "create_workspace",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "type": "null"
                  },
                  {
                    "$ref": "#/components/schemas/CreateWorkspaceRequest"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreatedWorkspaceResponse"
                }
              }
            }
          },
          "401": {
            "description": "No credential, or one this route does not accept"
          },
          "403": {
            "description": "The credential is good and lacks the capability this route requires"
          },
          "409": {
            "description": "The request conflicts with the resource's current state"
          },
          "500": {
            "description": "The daemon failed to answer"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateWorkspaceRequest": {
        "type": "object",
        "description": "`POST /v1/workspaces` — create one.",
        "properties": {
          "name": {
            "type": [
              "string",
              "null"
            ],
            "description": "What the workspace will be called, when the caller cares."
          }
        }
      },
      "CreatedWorkspaceResponse": {
        "type": "object",
        "description": "What creating answers with.",
        "required": [
          "workspace_id",
          "name",
          "request_id",
          "tenant_id"
        ],
        "properties": {
          "name": {
            "type": "string",
            "description": "Its name — echoed when chosen, revealed when generated."
          },
          "request_id": {
            "type": "string",
            "description": "The correlation token, in the body as `lifecycle.zig` writes it."
          },
          "tenant_id": {
            "type": "string",
            "description": "The tenant it was created in — the daemon's resolution, never a claim."
          },
          "workspace_id": {
            "type": "string",
            "description": "The new workspace's identifier."
          }
        }
      },
...
      "WorkspaceSummary": {
        "type": "object",
        "description": "One workspace as the list shows it.",
        "required": [
          "id",
          "created_at"
        ],
        "properties": {
          "created_at": {
            "type": "integer",
            "format": "int64",
            "description": "When it was created; the walk's sort key."
          },
          "id": {
            "type": "string",
            "description": "The workspace's identifier."
          },
          "name": {
            "type": [
              "string",
              "null"
            ],
            "description": "Its name — `null` on rows older than naming, emitted either way for\nthe tenant module's null rule."
          }
        }
      },
      "WorkspacesResponse": {
        "type": "object",
        "description": "`GET /v1/tenants/me/workspaces` — one page of the tenant's workspaces.",
        "required": [
          "items",
          "tenant_id"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WorkspaceSummary"
            },
            "description": "The rows on this page, oldest first."
          },
          "next_cursor": {
            "type": [
              "string",
              "null"
            ],
            "description": "Where the next page resumes, or `null` on the last page."
          },
          "tenant_id": {
            "type": "string",
            "description": "Whose list this is — the authoritative resolution, carried so a client\ncan pin its local state to the right tenant."
          },
          "total": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "Always `null`: the walk never counts, and the key stays because a\nclient may already branch on its presence."
          }
        }
      }
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Obtain a token via the CLI auth flow (POST /v1/auth/sessions) or GitHub OAuth"
      }
    }
  }
}